moved placeholder and validation code to different layers of type hierarchy
