+++ src/main/java/org/openhds/service/contract/AbstractAuditableService.java
@@ -2,8 +2,11 @@ package org.openhds.service.contract;
 
 import org.openhds.domain.contract.AuditableEntity;
 import org.openhds.domain.model.census.LocationHierarchy;
+import org.openhds.errors.model.Error;
 import org.openhds.errors.model.ErrorLog;
 import org.openhds.errors.model.ErrorLogException;
+import org.openhds.errors.util.ErrorLogger;
+import org.openhds.events.model.Event;
 import org.openhds.repository.contract.AuditableRepository;
 import org.openhds.repository.queries.QueryRange;
 import org.openhds.repository.results.EntityIterator;
@@ -15,7 +18,10 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 
+import javax.validation.ConstraintViolation;
+import javax.validation.Validator;
 import java.time.ZonedDateTime;
+import java.util.List;
 import java.util.Set;
 
 /**
@@ -35,13 +41,46 @@ public abstract class AbstractAuditableService
     @Autowired
     private UserHelper userHelper;
 
+    public T makePlaceHolder(String id) {
+        return makePlaceHolder(id, AuditableEntity.PLACEHOLDER_STATUS);
+    }
+
+    public T findOrMakePlaceHolder(String id){
+        if (exists(id)) {
+            return findOne(id);
+        }
+
+        T entity = makePlaceHolder(id);
+        entity.setUuid(id);
+
+        log.info("Making placeholder: " + entity);
+
+        return createOrUpdate(entity);
+    }
+
     @Override
     public T createOrUpdate(T entity) {
 
         checkNonStaleModifiedDate(entity);
         setAuditableFields(entity);
 
-        return super.createOrUpdate(entity);
+        ErrorLog errorLog = new ErrorLog();
+        errorLog.setEntityType(entity.getClass().getSimpleName());
+        verify(entity, errorLog);
+
+        String entityStatus = entity.getEntityStatus();
+        if(entityStatus.equals(entity.NORMAL_STATUS) || entityStatus.equals(entity.NEW_STATUS)) {
+            validate(entity, errorLog);
+        }
+
+        if (!errorLog.getErrors().isEmpty()) {
+            errorLogger.log(errorLog);
+            throw new ErrorLogException(errorLog);
+        }
+
+        T saved = repository.save(entity);
+        logEvent(saved);
+        return saved;
     }
 
     @Override
@@ -157,9 +196,8 @@ public abstract class AbstractAuditableService
         return iteratorFromPageable(pageable -> repository.findByDeletedFalseAndVoidBy(user, pageable), sort);
     }
 
-    @Override
     public void validate(T entity, ErrorLog errorLog) {
-        super.validate(entity, errorLog);
+
     }
 
     public EntityIterator<T> findByLastModifiedBy(Sort sort, User user) {
